delete property pop-up

+++ src/main/webapp/WEB-INF/view/landlord.jsp
@@ -83,11 +83,33 @@
                 <b>Carparks: </b>${house.carparks}<br>
                 <br>
                 <a class="btn btn-danger"
-                   href="/delete_house?houseId=${house.id}"
-                   onclick="return confirm('Delete this house?')">Delete</a>
+                   href="#"
+                   data-toggle="modal" data-target="#housemodal_${house.id}">Delete</a>
             </div>
         </div>
         <br>
+        <!-- Modal -->
+        <div class="modal fade" id="housemodal_${house.id}" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
+             aria-hidden="true">
+            <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="exampleModalLabel">Housing</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    <div class="modal-body">
+                        Delete this property?
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
+                        <a class="btn btn-danger"
+                           href="/delete_house?houseId=${house.id}">Confirm</a>
+                    </div>
+                </div>
+            </div>
+        </div>
     </c:forEach>
